Adds ability to pull images from resource folder without intervention.
Images are now automatically "gathered" starting from a root specified
in application.properties. This is transposed to a path relative to a
second ("target") root, then the list is fed into the existing
ImageCollection. Note that image display still needs lots of work; this
is just a step on the backend.