Resolves issue with WAR-run code on AWS not finding files.
Image files weren't in the right location to be picked up for inclusion
in the WAR generated by Maven. For some reason, this only became a
problem while configuring HTTPS. Note that the old path was correct
relative to source, just not relative to the WAR.

Additionally, the old resume path became a problem for some reason, in
the same manner. Corrected.

Tests still need to use the source location, and have been updated to
show that.

Additional refactorings; see per-file updates.