Windows image-path issue: Fixed, refactored problem areas.
Attempts to load art files from art.jsp were failing; instead of a
relative path, in my Windows dev environment, images were being created
with Image.relativeResourcePath set to the Windows canonical path
instead of a path which could be retrieved by the webapp. Attempts to
find the issue led to refactors and fixes across most areas touched by
file loading, with the MavenDefeatingFileLoader receiving the bulk of
changes, and greater test coverage.

Note that this includes a currently-unused method,
MavenDefeatingFileLoader.getFolderResource(). It was part of an
intermediate solution, and as it's already received test coverage and
might be useful in future, I've kept it.